Mesh points with real scale (nanometers)
I was not able to load the real scale from the TIF file in Paraview. For that reason I had to apply a Transform filter to force the mesh to have real scale (nanometer). We have to make sure to use the same value of `pixel_size` in both StackGen and SurfaceGen routines.